Refactor for merging
