+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import PastEventPage from './pastEventPage'
+import PastEventPage from './pastEvent/pastEventPage'
 import "./assets/css/main.css"
 
 ReactDOM.createRoot(document.getElementById('root')).render(
